requirements.txt
-e .

clusteval
fuzzy-c-means
graphviz
h5py
hdbscan
ipython
ipywidgets
joblib
jupyterlab
jupyterlab-widgets
librosa
matplotlib==3.6.0
nara_wpe
natsort
nltk
noisereduce
numpy
openpyxl
opencv-python
pandas
pillow
plotly
pydub
# pygraphviz; sys_platform == 'linux'
pymupdf
SciencePlots
scikit-image[optional]
scikit-learn
scipy
seaborn
SoundFile
sphinx
sphinx-rtd-theme
tables
tensorflow
tqdm
typing-extensions
# umap-learn
umap-learn[plot]
# voila
